Bluetooth: hci_bcm: Set the init baudrate too on apple.
The device must be accessed via the configured baudrate for all
operations, or the operations will time out.

Note that currently bcm_set_baudrate() always fails (the commands always
return -EBUSY) and hence the operating baudrate is never actually set.
But if it were, it would need to be the same as the init baudrate, so
we keep setting oper_speed anyway.

Signed-off-by: Ronald Tschalär <ronald@innovation.ch>