Bluetooth: hci_bcm: Add helper to toggle device wake GPIO
There is already a bcm_gpio_set_power() helper to toggle the power
enable pin.  Add a corresponding helper to toggle the device wake pin.
This will allow us to more easily integrate support for MacBooks which
use custom ACPI methods to access both pins.

The only functional change here is an additional 15 ms delay when
toggling the device wake pin in bcm_gpio_set_power().  It is unclear why
this delay is observed in bcm_suspend_device() and bcm_resume_device(),
but not in bcm_gpio_set_power() so far.

Signed-off-by: Lukas Wunner <lukas@wunner.de>