mfd: apple-ibridge: Add Apple iBridge MFD driver.
The iBridge chip provides access to several devices, including:
- the Touch Bar
- the iSight webcam
- the light sensor

This driver provides the core support for managing the iBridge and
the access to the devices. In particular, since the functionality for
the touch bar and light sensor is exposed via USB HID interfaces, and the
same HID device is used for multiple functions, this driver provides a
multiplexing layer that allows multiple HID drivers to be registered for
a given HID device. This allows the touch bar and ALS driver to separated
out into their own modules.

Signed-off-by: Ronald Tschalär <ronald@innovation.ch>